+++ Sources/KeyboardKit/Autocomplete/AutocompleteContext+Settings.swift
@@ -13,6 +13,19 @@ public extension AutocompleteContext {
 
     /// This type is used for autocomplete-related settings.
     ///
+    /// The ``isNextWordPredictionEnabled`` setting is false
+    /// by default, since users must explicitly approve that
+    /// their typed text is sent to a 3rd party service.
+    ///
+    /// You can use the ``nextWordPredictionRequestType`` to
+    /// let your users select which request type to use, and
+    /// ``nextWordPredictionRequestApiKey`` to define an API
+    /// key, for when you don't want to use your own API key.
+    ///
+    /// KeyboardKit Pro's ``KeyboardApp/SettingsScreen`` has
+    /// an visibility setting to enable a separate next word
+    /// prediction section, which by default is hidden.
+    ///
     /// All properties in this type are automatically stored
     /// in ``Foundation/UserDefaults/keyboardSettings`` with
     /// an `autocomplete` prefix.
@@ -49,6 +62,14 @@ public extension AutocompleteContext {
         /// Whether next word prediction is enabled, by default `true`.
         @AppStorage("\(settingsPrefix)isNextWordPredictionEnabled", store: .keyboardSettings)
         public var isNextWordPredictionEnabled = true
+
+        /// A custom, user-specified next word predicton request API key.
+        @AppStorage("\(settingsPrefix)nextWordPredictionApiKey", store: .keyboardSettings)
+        public var nextWordPredictionRequestApiKey = ""
+
+        /// A custom, user-specified next word predicton request type.
+        @AppStorage("\(settingsPrefix)nextWordPredictionRequestType", store: .keyboardSettings)
+        public var nextWordPredictionRequestType = Autocomplete.NextWordPredictionRequestType.claude
 
         /// The number of autocomplete suggestions to display, by default `3`.
         @AppStorage("\(settingsPrefix)suggestionsDisplayCount", store: .keyboardSettings)